Authentication is Done - each user manage their own contacts

+++ controllers/contactController.js
@@ -49,6 +49,11 @@ const updateContact = asyncHandler(async (req, res) => {
     res.status(404);
     throw new Error("Contact not found");
   }
+  if (contact.user_id.toString() !== req.user.id) {
+    // different user is trying to update a contact of another user
+    res.status(403);
+    throw new Error("User don't have permission to update other user contacts");
+  }
   const updatedContact = await Contact.findByIdAndUpdate(
     req.params.id,
     req.body,
@@ -65,6 +70,11 @@ const deleteContact = asyncHandler(async (req, res) => {
   if (!contact) {
     res.status(404);
     throw new Error("Contact not found");
+  }
+  if (contact.user_id.toString() !== req.user.id) {
+    // different user is trying to delete a contact of another user
+    res.status(403);
+    throw new Error("User don't have permission to delete other user contacts");
   }
   await Contact.deleteOne({ _id: req.params.id });
   res.status(200).json(contact);
